requirements.txt
# Core dependencies
python-dotenv
openai
httpx
sentence-transformers
qdrant-client
networkx
matplotlib
numpy
tinydb
beautifulsoup4
requests
tiktoken
trafilatura

# Server
fastapi
uvicorn[standard]
websockets
pyyaml

# Optional adapters
python-telegram-bot  # only needed if telegram adapter is enabled
